Fix IE Manager and Selenium Manager
