+++ src/main/java/io/github/bonigarcia/wdm/managers/ChromeDriverManager.java
@@ -30,6 +30,7 @@ import java.util.Optional;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import io.github.bonigarcia.wdm.config.DriverManagerType;
+import io.github.bonigarcia.wdm.online.S3BucketListNamespaceContext;
 
 import javax.xml.namespace.NamespaceContext;
 
@@ -153,27 +154,6 @@ public class ChromeDriverManager extends WebDriverManager {
     @Override
     protected NamespaceContext getNamespaceContext() {
         return S3_BUCKET_LIST_NAMESPACE_CONTEXT;
-    }
-
-    private static final class S3BucketListNamespaceContext implements NamespaceContext {
-
-        private static final String S3_BUCKET_LIST_NS = "http://doc.s3.amazonaws.com/2006-03-01";
-
-        @Override
-        public String getNamespaceURI(String prefix) {
-            return S3_BUCKET_LIST_NS;
-        }
-
-        @Override
-        public String getPrefix(String namespaceURI) {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public Iterator getPrefixes(String namespaceURI) {
-            throw new UnsupportedOperationException();
-        }
-
     }
 
 }
